requirements.txt
argparse
glob2
psycopg2-binary
python-dotenv